Refactored condition controls, started adding drag and drop features
