+++ examples/pages/ConditionSelector.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta http-equiv="x-ua-compatible" content="ie=edge">
+    <title>ConditionSelector</title>  
+  </head>
+  <body>
+    <div id="root"></div>    
+    <script src="ConditionSelector.js"></script>
+  </body>
+</html>
