+++ examples/src/Dropdown.tsx
@@ -6,176 +6,17 @@ import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 
 import { Dropdown } from './../../src/components/Dropdown/Dropdown';
-import { LogicalOperatorTypeEnum, PropertyTypeEnum, ExpressionDefinitionTree, ConditionDefinition } from './../../src/components/ConditionSelector/ConditionDefinitionRow.Props';
-import { ConditionSelector } from './../../src/components/ConditionSelector/ConditionSelector';
 import { DropdownType } from './../../src/components/Dropdown/Dropdown.Props';
-
 import { Label } from './../../src/components/Label/Label';
 import { Icon } from './../../src/components/Icon/Icon';
 import { Slider } from './../../src/components/Slider/Slider';
 import { Button } from './../../src/components/Button/Button';
-import { TextField } from './../../src/components/TextField/TextField';
 
 export class Index extends React.Component<any, any> {
     
     public render() {
-        const treeviewElements = [{id: '1', text: 'Script returns any data'}, {id: '2', text: 'Script does not return data'}, {id: '3', text: 'Exception occurs while executing the script'}];
-        const treeviewElements2 = [{id: '1', text: 'With specific Name'}, {id: '2', text: 'With specific Status'}, {id: '3', text: 'With specific Restarted On'}, {id: '4', text: 'Mirko'}];
-        const conditionElements = [{
-            id: 3,
-            isHardcodedValue: true,
-            propertyName: 'Script returns any data',
-            hasMultipleLogicalOperations: false,
-            allowConditionDeletion: false,
-            propertyType: PropertyTypeEnum.None
-        }, {
-            id: 2,
-            isHardcodedValue: false,
-            propertyName: 'Network ID',
-            hasMultipleLogicalOperations: false,
-            conditionSelectionTypes:  [{ key: 'And', text: 'Equals' }, { key: 'Or', text: 'Less than', selected: true }],
-            propertyType: PropertyTypeEnum.Number
-        }, {
-            id: 25,
-            isHardcodedValue: false,
-            propertyName: 'Network ID',
-            hasMultipleLogicalOperations: false,
-            conditionSelectionTypes:  [{ key: 'And', text: 'Equals' }, { key: 'Or', text: 'More than ', selected: true }],
-            propertyType: PropertyTypeEnum.Number,
-            hasIndent: true
-        }, {
-            id: 3,
-            isHardcodedValue: false,
-            propertyName: 'Name',
-            conditionSelectionTypes: [{ key: 'Is', text: 'Is'}, { key: 'Or', text: 'Starts with', selected: true }, { key: 'Ise', text: 'I don\'t care'}],
-            hasIndent: false,
-            propertyType: PropertyTypeEnum.String
-        }, {
-            id: 4,
-            isHardcodedValue: false,
-            propertyName: 'Running',
-            conditionSelectionTypes: [{ key: 'Is', text: 'Is'}, { key: 'Or', text: 'Is not', selected: true}],
-            propertyType: PropertyTypeEnum.Boolean
-        }, {
-            id: 5,
-            isHardcodedValue: false,
-            propertyName: 'Status',
-            conditionSelectionTypes: [{ key: 'Is', text: 'Is', selected: true}, { key: 'Or', text: 'Is not'}],
-            additionalData: [{ key: 'Is', text: 'Running', selected: true}, { key: 'Or', text: 'Stopped'}, { key: 'O1r', text: 'Restarting'}, { key: 'Or2', text: 'Restart pending'}],
-            propertyType: PropertyTypeEnum.Enum
-        }, {
-            id: 6,
-            isHardcodedValue: false,
-            propertyName: 'Restarted on',
-            conditionSelectionTypes: [{ key: 'Is', text: 'On', selected: true}, { key: 'O1r', text: 'Before'}, { key: 'O2r', text: 'After'}],
-            propertyType: PropertyTypeEnum.DateTime
-        }];
-        let conditionElements2: ExpressionDefinitionTree = {logicalOperator: LogicalOperatorTypeEnum.None, 
-                                                            subExpressions: [{ logicalOperator: LogicalOperatorTypeEnum.None,
-                                                                                conditionDefinition: {
-                                                                                    id: 1,
-                                                                                    isHardcodedValue: true,
-                                                                                    propertyName: 'Script returns any data',
-                                                                                    hasMultipleLogicalOperations: false,
-                                                                                    allowConditionDeletion: false,
-                                                                                    propertyType: PropertyTypeEnum.None
-                                                                                 }}, 
-                                                                                {logicalOperator: LogicalOperatorTypeEnum.And,
-                                                                                conditionDefinition: {
-                                                                                    id: 2,
-                                                                                    isHardcodedValue: false,
-                                                                                    propertyName: 'Network ID',
-                                                                                    hasMultipleLogicalOperations: false,
-                                                                                    conditionSelectionTypes:  [{ key: 'And', text: 'Equals' }, { key: 'Or', text: 'Less than', selected: true }],
-                                                                                    propertyType: PropertyTypeEnum.Number
-                                                                                }},
-                                                                            {logicalOperator: LogicalOperatorTypeEnum.Or,
-                                                                                subExpressions: 
-                                                                                    [
-                                                                                        {logicalOperator: LogicalOperatorTypeEnum.None,
-                                                                                            conditionDefinition: {
-                                                                                                id: 3,
-                                                                                                isHardcodedValue: false,
-                                                                                                propertyName: 'Name',
-                                                                                                conditionSelectionTypes: [{ key: 'Is', text: 'Is'}, { key: 'Or', text: 'Starts with', selected: true }, { key: 'Ise', text: 'I don\'t care'}],                                                                                                
-                                                                                                propertyType: PropertyTypeEnum.String
-                                                                                            }
-                                                                                        },
-                                                                                        {logicalOperator: LogicalOperatorTypeEnum.And,
-                                                                                            conditionDefinition: {
-                                                                                                id: 4,
-                                                                                                isHardcodedValue: false,
-                                                                                                propertyName: 'Running',
-                                                                                                conditionSelectionTypes: [{ key: 'Is', text: 'Is'}, { key: 'Or', text: 'Is not', selected: true}],
-                                                                                                propertyType: PropertyTypeEnum.Boolean
-                                                                                            }
-                                                                                        }
-                                                                                    ]
-                                                                            },
-                                                                            {logicalOperator: LogicalOperatorTypeEnum.And,
-                                                                                conditionDefinition: {
-                                                                                    id: 5,
-                                                                                    isHardcodedValue: false,
-                                                                                    propertyName: 'Status',
-                                                                                    conditionSelectionTypes: [{ key: 'Is', text: 'Is', selected: true}, { key: 'Or', text: 'Is not'}],
-                                                                                    additionalData: [{ key: 'Is', text: 'Running', selected: true}, { key: 'Or', text: 'Stopped'}, { key: 'O1r', text: 'Restarting'}, { key: 'Or2', text: 'Restart pending'}],
-                                                                                    propertyType: PropertyTypeEnum.Enum
-                                                                                }
-                                                                            },
-                                                                            {
-                                                                                logicalOperator: LogicalOperatorTypeEnum.And,
-                                                                                subExpressions: [
-                                                                                    {
-                                                                                        logicalOperator: LogicalOperatorTypeEnum.Null,
-                                                                                        subExpressions: [
-                                                                                            {
-                                                                                                logicalOperator: LogicalOperatorTypeEnum.None,
-                                                                                                conditionDefinition: {
-                                                                                                    id: 6,
-                                                                                                    isHardcodedValue: false,
-                                                                                                    propertyName: 'Restarted on',
-                                                                                                    conditionSelectionTypes: [{ key: 'Is', text: 'On', selected: true}, { key: 'O1r', text: 'Before'}, { key: 'O2r', text: 'After'}],
-                                                                                                    propertyType: PropertyTypeEnum.DateTime
-                                                                                                }
-                                                                                            },
-                                                                                            {
-                                                                                                logicalOperator: LogicalOperatorTypeEnum.Or,
-                                                                                                conditionDefinition: {
-                                                                                                    id: 7,
-                                                                                                    isHardcodedValue: false,
-                                                                                                    propertyName: 'Display Name',
-                                                                                                    conditionSelectionTypes: [{ key: 'Is', text: 'Is'}, { key: 'Or', text: 'Starts with', selected: true }, { key: 'Ise', text: 'I don\'t care'}],                                                                                                
-                                                                                                    propertyType: PropertyTypeEnum.String
-                                                                                                }                                                                                            
-                                                                                            }
-                                                                                        ]
-                                                                                    },
-                                                                                    {
-                                                                                        logicalOperator: LogicalOperatorTypeEnum.And,
-                                                                                        conditionDefinition: {
-                                                                                            id: 2,
-                                                                                            isHardcodedValue: false,
-                                                                                            propertyName: 'Sex appeal',
-                                                                                            hasMultipleLogicalOperations: false,
-                                                                                            conditionSelectionTypes:  [{ key: 'And', text: 'Equals' }, { key: 'Or', text: 'Less than', selected: true }],
-                                                                                            propertyType: PropertyTypeEnum.Number
-                                                                                        }  
-                                                                                    }
-                                                                                ]
-                                                                            }
-                                                                            ]};
         return (
             <div>
-
-                    <ConditionSelector 
-                        specialConditionsList = {treeviewElements}
-                        standardConditionsList = {treeviewElements2}
-                        selectedConditions= {conditionElements2}
-                    />
-
-
-
-
                 <Dropdown
                     hasTitleBorder={true}
                     dropdownType={DropdownType.selectionDropdown}
