+++ package.json
@@ -29,8 +29,8 @@
         "node-sass": "4.5.3",
         "react": "15.6.1",
         "react-datetime": "2.8.9",
-        "react-dnd": "2.4.0",
-        "react-dnd-html5-backend": "2.4.1",
+        "react-dnd": "^2.4.0",
+        "react-dnd-html5-backend": "^2.4.1",
         "react-dom": "15.6.1",
         "react-draggable": "2.2.3",
         "react-numeric-input": "2.0.7",
