+++ src/components/ConditionSelector/Condition.scss
@@ -1,47 +1,19 @@
 @import '../../index.scss';
-.condition-definition-item {
-    .condition-definition-item {
-        display: flex;
-        justify-content: space-between;
-        margin: 10px 5px;
-        .conditions-group-container {
-            width: 100%;
-            .condition-definition-item {
-                margin-right: 0;
-            }
-            .condition-definition-item:first-child  {
-                margin-top: 0;
-            }
-            .condition-definition-item:last-child  {
-                margin-bottom: 0;
-            }
-        }
-        .single-condition-container {
-            width: 100%;            
-            &:hover {
-                cursor: move;
-            }
-        }
-        .condition-operator {
-            border: 1px solid transparent;
-            background-color: #F0F0F0;
-            border-radius: 8px;
-            padding: 5px 3px;
-            margin-right: 3px;
-            width: 50px;
-        }
-    }
-}
+
 .command-definition-row-container {
     display: flex;
     justify-content: space-between;
     align-items: center;
     border: 1px solid transparent;
-    background-color: #F0F0F0;
     border-radius: 8px;
+    background-color: #F0F0F0;
     padding: 5px;
     height: 40px;
     
+    &.command-in-drag {
+        opacity: 0.3;
+    }
+
     .left-content {
         display: flex;
         align-items: center;
@@ -56,6 +28,7 @@
             padding-right: 15px;
         }
     }
+
     .dropdown, .text-field {
         margin-bottom: 0;
     }
