+++ src/components/ConditionSelector/ConditionGroup.scss
@@ -1,0 +1,45 @@
+@import '../../index.scss';
+ 
+.condition-definition-item {
+    .condition-definition-item {
+        display: flex;
+        justify-content: space-between;
+        margin: 10px 5px;
+        .conditions-group-container {
+            width: 100%;
+            .condition-definition-item {
+                margin-right: 0;
+            }
+            .condition-definition-item:first-child  {
+                margin-top: 0;
+            }
+            .condition-definition-item:last-child  {
+                margin-bottom: 0;
+            }
+        }
+        .single-condition-container {
+            width: 100%;            
+            &:hover, &:active {
+                cursor: move;
+            }            
+            .droppable{
+                height: 40px;
+                width: 100%;
+                border: 1px solid #1E88E5;
+                border-radius: 8px;
+            }
+
+            .hiddenDroppable{
+                display: none;
+            }
+        }
+        .condition-operator {
+            border: 1px solid transparent;
+            background-color: #F0F0F0;
+            border-radius: 8px;
+            padding: 5px 3px;
+            margin-right: 3px;
+            width: 50px;
+        }
+    }
+}
